Move check to accept auth; Remove old benchmark
